refactor(server): split handler.go by route family
Pure move: handler.go went from 2851 lines / 79 funcs to eleven files:

  handler.go            475  core: type, NewHandler, Router, middleware
  handler_phones.go     614  /phones routes + line settings
  handler_linkhealth.go 391  /api/calls/{id}/link_health + SSE
  handler_dashboard.go  285  /, /connecting, dashboard helpers
  handler_links.go      228  /links routes (household links)
  handler_ws.go         220  /ws + dev harness endpoints
  handler_helpers.go    207  ownership, household, render helpers
  handler_api.go        176  /api/* JSON routes
  handler_settings.go   165  /settings routes
  handler_calls.go      152  /calls + /call/live/{id}
  handler_onboard.go     63  /onboard

No logic change; goimports prunes each file's imports. Tests pass.
